tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/script-defer.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: HTMLScriptElement - defer</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

let script_run_status = "inline";
let t = async_test("the defer script run later");

</script>

<script type="text/javascript" src="defer.js" defer></script>

<script>

t.step(() => {
  assert_equals(script_run_status, "inline", "the script run status");
  script_run_status = "deferred";
});

</script>
